Refactor directory structure for clarity and improved organization
This commit reorganizes the project directory structure to enhance clarity and improve organization. The changes include:

- **Services Separation**: Services are now organized into domain-specific subdirectories within the `applications/services` directory. This separation into `admin` and `tax` subdirectories isolates domain logic and enhances the manageability of service-related code.
