+++ applications/services/admin/admin_service.go
@@ -1,4 +1,4 @@
-package services
+package admin
 
 type AdminServiceInterface interface {
 	UpdatePersonalDeduction(amount float64) error
