+++ applications/services/admin/admin_service_impl.go
@@ -1,4 +1,4 @@
-package services
+package admin
 
 import (
 	"errors"
